requirements.txt
gTTS
openai-whisper
pydub
